refactor: move relay files into own dir
